+++ vmware_horizon.py
@@ -2721,7 +2721,10 @@ class Federation:
         Available for Horizon 8 2012 and later."""
         response = requests.get(f'{self.url}/rest/federation/v1/sites', verify=False,  headers=self.access_token)
         if response.status_code == 400:
-            error_message = (response.json())["error_message"]
+            if "error_messages" in response.json():
+                error_message = (response.json())["error_messages"]
+            else:
+                error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         if response.status_code == 404:
             error_message = (response.json())["error_message"]
@@ -2744,7 +2747,10 @@ class Federation:
         Available for Horizon 8 2012 and later."""
         response = requests.get(f'{self.url}/rest/federation/v1/sites/{site_id}', verify=False,  headers=self.access_token)
         if response.status_code == 400:
-            error_message = (response.json())["error_message"]
+            if "error_messages" in response.json():
+                error_message = (response.json())["error_messages"]
+            else:
+                error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         if response.status_code == 404:
             error_message = (response.json())["error_message"]
@@ -2773,7 +2779,10 @@ class Federation:
         json_data = json.dumps(data)
         response = requests.put(f'{self.url}/rest/federation/v1/cpa', verify=False,  headers=headers, data = json_data)
         if response.status_code == 400:
-            error_message = (response.json())["error_message"]
+            if "error_messages" in response.json():
+                error_message = (response.json())["error_messages"]
+            else:
+                error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         if response.status_code == 404:
             error_message = (response.json())["error_message"]
@@ -2795,7 +2804,10 @@ class Federation:
         Available for Horizon 8 2012 and later."""
         response = requests.get(f'{self.url}/rest/federation/v1/home-sites/1/{home_site_id}', verify=False,  headers=self.access_token)
         if response.status_code == 400:
-            error_message = (response.json())["error_message"]
+            if "error_messages" in response.json():
+                error_message = (response.json())["error_messages"]
+            else:
+                error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         if response.status_code == 404:
             error_message = (response.json())["error_message"]
@@ -2825,7 +2837,10 @@ class Federation:
         json_data = json.dumps(data)
         response = requests.post(f'{self.url}/rest/federation/v1/sites', verify=False,  headers=headers, data = json_data)
         if response.status_code == 400:
-            error_message = (response.json())["error_message"]
+            if "error_messages" in response.json():
+                error_message = (response.json())["error_messages"]
+            else:
+                error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         if response.status_code == 404:
             error_message = (response.json())["error_message"]
@@ -2853,7 +2868,10 @@ class Federation:
         json_data = json.dumps(data)
         response = requests.put(f'{self.url}/rest/federation/v1/sites/{site_id}', verify=False,  headers=headers, data = json_data)
         if response.status_code == 400:
-            error_message = (response.json())["error_message"]
+            if "error_messages" in response.json():
+                error_message = (response.json())["error_messages"]
+            else:
+                error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         if response.status_code == 404:
             error_message = (response.json())["error_message"]
@@ -2874,7 +2892,10 @@ class Federation:
         Available for Horizon 8 2012 and later."""
         response = requests.delete(f'{self.url}/rest/federation/v1/sites/{site_id}', verify=False,  headers=self.access_token)
         if response.status_code == 400:
-            error_message = (response.json())["error_message"]
+            if "error_messages" in response.json():
+                error_message = (response.json())["error_messages"]
+            else:
+                error_message = (response.json())["error_message"]
             raise Exception(f"Error {response.status_code}: {error_message}")
         if response.status_code == 404:
             error_message = (response.json())["error_message"]
@@ -2890,3 +2911,91 @@ class Federation:
                 raise "Error: " + str(e)
             else:
                 return response.json()
+
+    def get_pods(self) -> list:
+        """Lists all the pods in the pod federation.
+
+        Available for Horizon 8 2012 and later."""
+        response = requests.get(f'{self.url}/rest/federation/v1/pods', verify=False,  headers=self.access_token)
+        if response.status_code == 400:
+            if "error_messages" in response.json():
+                error_message = (response.json())["error_messages"]
+            else:
+                error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        if response.status_code == 404:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 403:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+    def get_pod(self, pod_id:str) -> dict:
+        """Retrieves a given pod from the pod federation.
+
+        Available for Horizon 8 2012 and later."""
+        response = requests.get(f'{self.url}/rest/federation/v1/pods/{pod_id}', verify=False,  headers=self.access_token)
+        if response.status_code == 400:
+            if "error_messages" in response.json():
+                error_message = (response.json())["error_messages"]
+            else:
+                error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        if response.status_code == 404:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 403:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        elif response.status_code != 200:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+            else:
+                return response.json()
+
+    def update_pod(self,pod_id:str, description:str, site_id:str ,name:str, cloud_managed:bool="") -> dict:
+        """Updates the given pod in the pod federation.
+
+        Requires pod_id, site_id, the name and description as strings, cloud_managed needs to be a bool
+        Available for Horizon 8 2012 and later."""
+        headers = self.access_token
+        headers["Content-Type"] = 'application/json'
+        data = {}
+        if cloud_managed != "":
+            data["cloud_managed"] = cloud_managed
+        if description!="":
+            data["description"] = description
+        data["name"] = name
+        data["site_id"] = site_id
+        json_data = json.dumps(data)
+        response = requests.put(f'{self.url}/rest/federation/v1/pods/{pod_id}', verify=False,  headers=headers, data = json_data)
+        if response.status_code == 400:
+            if "error_messages" in response.json():
+                error_message = (response.json())["error_messages"]
+            else:
+                error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        if response.status_code == 404:
+            error_message = (response.json())["error_message"]
+            raise Exception(f"Error {response.status_code}: {error_message}")
+        elif response.status_code == 403:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        elif response.status_code != 204:
+            raise Exception(f"Error {response.status_code}: {response.reason}")
+        else:
+            try:
+                response.raise_for_status()
+            except requests.exceptions.RequestException as e:
+                raise "Error: " + str(e)
+