+++ changelog.md
@@ -212,3 +212,9 @@
     * Added update_site
     * Added delete_site
     * Added new_site
+
+10-03-2021
+* federation class
+    * Added get_pods
+    * Added get_pod
+    * Added update_pod